few more updates
