+++ packages/web/.eslintignore
@@ -1,1 +1,2 @@
 dist
+index.html
