+++ packages/web/index.html
@@ -4,8 +4,8 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#7ee2a8" />
-    <link rel="icon" href="favicon.svg" />
-  <title>Tradingview Shariah Indicator - TSI</title>
+    <link rel="icon" href="/favicon.svg" />
+    <title>Tradingview Shariah Indicator - TSI</title>
     <meta name="description" content="Add a small green indicator in tradingview.com. At the moment, only cover Malaysian, NYSE, Nasdaq, Shenzhen, Shanghai" />
 
     <meta property='og:title' content="Tradingview Shariah Indicator - TSI" />
