+++ packages/web/tsconfig.json
@@ -8,7 +8,7 @@
     "resolveJsonModule": true,
     "jsx": "preserve",
     "jsxImportSource": "solid-js",
-    "types": ["vite/client"],
+    "types": ["vite/client", "types.d.ts"],
     "baseUrl": "./",
     "paths": {
       "@util": ["src/util"],
