+++ packages/web/type.d.ts
@@ -1,0 +1,9 @@
+export type PageProps = {
+  data: [exchange: string, code: string, name: string][]
+  metadata: Record<string, Date>
+  exchangesList: string[]
+  queryParams: Partial<{
+    q: string
+    exchange: string | string[]
+  }>
+}
